split `ring` feature into aesgcm and default features
This commit splits the `ring` feature into to features:
 - `aesgcm`
 - `c20p1305`

By default `c20p1305` is enabled while users must opt-in
to `aesgcm` since c20p1305 should be more resistant against
side-channel attacks (timing) on hardware without AES-NI.

Therefore, all tests and examples move to `CHACHA20_POLY1305`
by default - but unit tests and benchmarks use `AES_256_GCM`
when the `aesgcm` feature is enabled.
